Toggle logging of ignored values in class method, not #define
This addresses the issue raised by @andrewsardone in
krzysztofzablocki/KZPropertyMapper#10, using his proposed
solution.

Users (CocoaPods users in particular) should now use the
class method `+[KZPropertyMapper logIgnoredValues:] to
configure whether to log ignored values.

The previous method, via `#define KZPropertyMapperLogIgnoredValues`,
will still work but is no longer exposed in the
public interface.

Credit to @andrewsardone for the idea and for suggesting
that the class method be placed in a Debug category,
to separate it from the class's main concern (that is,
mapping stuff).