[Tests] adds mapping tests for dependencies using protocols
When mapping anonymous objects that adhere to some protocol,
prefer using `id<SomeProtocol>` as the `property`'s type.

Avoid generic type definitions, such as `NSObject<SomeProtocol>`,
because type-encoding is more difficult to reason about. If a
concrete type declaration is publically available, such as a
`NSObject` subclass adhering to `SomeProtocol`, it is safe to use
the exact `class` type.